update binding config & scripts
